Se cambia el saldo de enteros a decimales

+++ src/com/kupay/ActualizarCC.java
@@ -79,7 +79,7 @@ class ActualizarCC extends AsyncTask<Void, Integer, JSONObject>{
      			
  				Log.v("app","Datos: "+ datos.toString());
  				TextView cc = (TextView) context.findViewById(R.id.cantidad);
- 				cc.setText("$"+Integer.toString(datos.getInt("SALDO")));
+ 				cc.setText("$"+Double.toString(datos.getDouble("SALDO")));
  				//AnimaSaldo actcc = new AnimaSaldo(context);
  				//actcc.equals(datos.getInt("SALDO"));
